Move io_connection_pattern to graph/op.py
