+++ pytensor/compile/builders.py
@@ -17,11 +17,10 @@ from pytensor.graph.basic import (
     NominalVariable,
     Variable,
     graph_inputs,
-    io_connection_pattern,
 )
 from pytensor.graph.fg import FunctionGraph
 from pytensor.graph.null_type import NullType
-from pytensor.graph.op import HasInnerGraph, Op
+from pytensor.graph.op import HasInnerGraph, Op, io_connection_pattern
 from pytensor.graph.replace import clone_replace
 from pytensor.graph.utils import MissingInputError
